refactor(srv): extract store.rs inline tests into store/tests submodule
