+++ .changesets/1782938765-refactor-srv-extract-store-rs-inline-tests-to-submodule-efys.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(srv): extract store.rs inline tests to submodule
